% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_summarize.R
\name{recursive_summarize}
\alias{recursive_summarize}
\title{Recursively summarize a dataset}
\usage{
recursive_summarize(dataset, groups, prompts, level = 1)
}
\arguments{
\item{dataset}{The dataset to be summarized}

\item{groups}{The variables by which the dataset should be grouped}

\item{prompts}{A set of strings used as prompts for each level of recursion}

\item{level}{The number of levels of recursion}
}
\value{
The input dataset with added columns for the summaries at
each level of recursion
}
\description{
Recursively summarize a dataset
}
