% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chunks.R
\name{split_chunks}
\alias{split_chunks}
\title{Split text into chunks}
\usage{
split_chunks(
  split,
  text,
  max_chars,
  max_chunks,
  split_on_newline = FALSE,
  gather_chunks = FALSE
)
}
\arguments{
\item{split}{logical indicating whether to split text into groups of max_chars (TRUE) or one chunk of max_chars (FALSE) # nolint}

\item{text}{character vector of text to be split}

\item{max_chars}{integer indicating the maximum number of characters per chunk}

\item{max_chunks}{integer indicating maximum number of chunks}

\item{split_on_newline}{boolean indicating whether to split on newline}

\item{gather_chunks}{boolean indicating whether to gather chunks together based on max_chars}
}
\value{
a list of character vectors representing the chunks of text
}
\description{
This function splits text into chunks of a specified maximum number of characters. If the split argument is set to TRUE, the text will be split into groups of the specified max_chars. If set to FALSE, the text will be split into one chunk of the specified max_chars.
}
\examples{
text <- "This is some text that I want to split into chunks."
split_chunks(TRUE, text, 10)
split_chunks(FALSE, text, 10)
This function takes in two parameters: str_list and max_chars.
It loops through each string in str_list and checks if the current string's length plus the total length of the previous strings is less than or equal to max_chars.
If it is, it adds the length of the current string to the total and increments the count of sequential strings. If it's not, it updates the maximum number of sequential strings and resets the total length and count of sequential strings. Finally, it returns the maximum number of sequential strings, either from the loop or the last iteration.
}
