% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_group.R
\name{split_group}
\alias{split_group}
\title{Split a string into groups of a given length}
\usage{
split_group(x, n)
}
\arguments{
\item{x}{A character string to split into groups}

\item{n}{The number of characters in each group}
}
\value{
A character vector containing the split groups
}
\description{
Split a string into groups of a given length
}
\examples{
split_group("abcdefghijklmnopqrstuvwxyz", 5)
#> "abcde" "fghij" "klmno" "pqrst" "uvwxy" "z"
}
